refactor: use constant for dms setting
