+++ lib/exchange_clients/bitfinex/index.js
@@ -14,6 +14,8 @@ const getMarkets = require('./get_markets')
 const unsubscribe = require('./unsubscribe')
 const subscribe = require('./subscribe')
 
+const { DMS_ENABLED } = require('../../constants')
+
 class BitfinexExchangeConnection {
   constructor (opts) {
     const { wsURL, restURL } = opts
@@ -31,7 +33,7 @@ class BitfinexExchangeConnection {
 
   setDMS (dms) {
     this.setAuthArgs({
-      dms: dms ? 4 : 0
+      dms: dms ? DMS_ENABLED : 0
     })
   }
