+++ lib/ws_servers/api/algos/algo_worker.js
@@ -6,7 +6,7 @@ const d = debug('bfx:hf:server:algo-worker')
 const { AOHost } = require('bfx-hf-algo')
 
 const validateAO = require('./util/validate_ao')
-const { DMS_ENABLED } = require('../constants')
+const { DMS_ENABLED } = require('../../../constants')
 
 class AlgoWorker {
   constructor (settings, algoOrders, bcast, algoDB) {
