+++ lib/ws_servers/api/handlers/on_save_api_credentials.js
@@ -14,7 +14,7 @@ const encryptAPICredentials = require('../../../util/encrypt_api_credentials')
 const isAuthorized = require('../../../util/ws/is_authorized')
 const capture = require('../../../capture')
 const openAuthBitfinexConnection = require('../open_auth_bitfinex_connection')
-const { DEFAULT_USER } = require('../constants')
+const { DEFAULT_USER } = require('../../../constants')
 
 module.exports = async (server, ws, msg) => {
   const { d, db, wsURL, restURL, reconnectAlgoHost } = server
