+++ lib/ws_servers/api/open_auth_bitfinex_connection.js
@@ -7,6 +7,7 @@ const { _default: DEFAULT_SETTINGS } = require('bfx-hf-ui-config').UserSettings
 const BitfinexExchangeConnection = require('../../exchange_clients/bitfinex')
 const { notifyInfo, notifyError, notifySuccess } = require('../../util/ws/notify')
 const send = require('../../util/ws/send')
+const { DMS_ENABLED } = require('../../constants')
 
 module.exports = async (conf) => {
   const { ws, apiKey, apiSecret, db, d, opts } = conf
@@ -21,7 +22,7 @@ module.exports = async (conf) => {
 
   d('opening auth bfx connection (dms %s)', dms ? 'enabled' : 'disabled')
 
-  client.setAuthArgs({ dms: dms ? 4 : 0 })
+  client.setAuthArgs({ dms: dms ? DMS_ENABLED : 0 })
   client.openWS({
     apiKey,
     apiSecret
