+++ lib/ws_servers/api/send_authenticated.js
@@ -5,7 +5,7 @@ const sendError = require('../../util/ws/send_error')
 const { notifySuccess } = require('../../util/ws/notify')
 const decryptAPICredentials = require('../../util/decrypt_api_credentials')
 const openAuthBitfinexConnection = require('./open_auth_bitfinex_connection')
-const { DEFAULT_USER } = require('./constants')
+const { DEFAULT_USER } = require('../../constants')
 
 module.exports = async (ws, db, d, opts) => {
   const { authPassword, authControl } = ws
